+++ src/collective/timestamp/browser/verification.py
@@ -2,7 +2,7 @@
 
 from collective.timestamp.interfaces import ITimeStamper
 from collective.timestamp.interfaces import ITimestampingSettings
-from collective.timestamp.utils import get_timestamp_date_from_tsr
+from collective.timestamp.utils import get_timestamp_date_from_tsr_file
 from plone import api
 from Products.Five.browser import BrowserView
 
@@ -15,7 +15,7 @@ class VerificationView(BrowserView):
 
     def timestamp_date(self):
         tsr_file = self.context.timestamp
-        timestamp_date = get_timestamp_date_from_tsr(tsr_file.data)
+        timestamp_date = get_timestamp_date_from_tsr_file(tsr_file.data)
         return timestamp_date.astimezone()
 
     def more_infos_url(self):